src/magnet/nets/coding_net/requirements.txt
# Requirements for coding_net ensemble system

# Core dependencies
pyyaml>=6.0
numpy>=1.24.0

# LangChain
langchain-core>=0.3.0
langchain-openai>=0.2.0

# OpenAI
openai>=1.0.0
